# fleet-pay-source-code 1 (1)/fleet-pay-export/openapi.yaml
openapi: '3.0.0'
info: 
  title: 'backend'
  version: '1.0.0'
...
paths: 
...
  /auth/changePassword: 
    post: 
      tags: 
        - 'Auth'
      operationId: 'changePassword'
      requestBody: 
        content: 
          application/json: 
            schema: 
              type: 'object'
              properties: 
                oldPassword: 
                  type: 'string'
                newPassword: 
                  type: 'string'
      responses: 
        200: 
          description: 'Success'
          content: 
            application/json: 
              schema: 
                type: 'object'
                properties: 
                  code: 
                    type: 'integer'
                    description: 'The status code of the response'
                  result: 
                    type: 'string'
                    nullable: false
                    description: ''
                  error: 
                    type: 'string'
                    nullable: true
                    description: 'Contains error message if any'
                required: 
                  - 'code'
  /auth/me: 
    get: 
      tags: 
        - 'Auth'
      operationId: 'me'
      responses: 
        200: 
          description: 'Success'
          content: 
            application/json: 
              schema: 
                type: 'object'
                properties: 
                  code: 
                    type: 'integer'
                    description: 'The status code of the response'
                  result: 
                    $ref: '#/components/schemas/User'
                    nullable: false
                    description: ''
                  error: 
                    type: 'string'
                    nullable: true
                    description: 'Contains error message if any'
                required: 
                  - 'code'
    patch: 
      tags: 
        - 'Auth'
      operationId: 'updateMe'
      requestBody: 
        content: 
          application/json: 
            schema: 
              type: 'object'
              properties: 
                user: 
                  $ref: '#/components/schemas/UserUpdateRequest'
      responses: 
        200: 
          description: 'Success'
          content: 
            application/json: 
              schema: 
                type: 'object'
                properties: 
                  code: 
                    type: 'integer'
                    description: 'The status code of the response'
                  result: 
                    $ref: '#/components/schemas/User'
                    nullable: false
                    description: ''
                  error: 
                    type: 'string'
                    nullable: true
                    description: 'Contains error message if any'
                required: 
                  - 'code'
  /auth/forgotPassword/sendOtp: 
    post: 
      tags: 
        - 'Auth'
      operationId: 'sendForgotPasswordOtp'
      requestBody: 
        content: 
          application/json: 
            schema: 
              type: 'object'
              properties: 
                email: 
                  type: 'string'
      responses: 
        200: 
          description: 'Success'
          content: 
            application/json: 
              schema: 
                type: 'object'
                properties: 
                  code: 
                    type: 'integer'
                    description: 'The status code of the response'
                  result: 
                    type: 'string'
                    nullable: false
                    description: ''
                  error: 
                    type: 'string'
                    nullable: true
                    description: 'Contains error message if any'
                required: 
                  - 'code'
  /auth/forgotPassword/validateOtp: 
    post: 
      tags: 
        - 'Auth'
      operationId: 'validateForgotPasswordOtp'
      requestBody: 
        content: 
          application/json: 
            schema: 
              type: 'object'
              properties: 
                requestId: 
                  type: 'string'
                token: 
                  type: 'string'
      responses: 
        200: 
          description: 'Success'
          content: 
            application/json: 
              schema: 
                type: 'object'
                properties: 
                  code: 
                    type: 'integer'
                    description: 'The status code of the response'
                  result: 
                    type: 'boolean'
                    nullable: false
                    description: ''
                  error: 
                    type: 'string'
                    nullable: true
                    description: 'Contains error message if any'
                required: 
                  - 'code'
  /auth/forgotPassword/changePassword: 
    post: 
      tags: 
        - 'Auth'
      operationId: 'changePasswordWithOtp'
      requestBody: 
        content: 
          application/json: 
            schema: 
              type: 'object'
              properties: 
                requestId: 
                  type: 'string'
                token: 
                  type: 'string'
                newPassword: 
                  type: 'string'
      responses: 
        200: 
          description: 'Success'
          content: 
            application/json: 
              schema: 
                type: 'object'
                properties: 
                  code: 
                    type: 'integer'
                    description: 'The status code of the response'
                  result: 
                    type: 'boolean'
                    nullable: false
                    description: ''
                  error: 
                    type: 'string'
                    nullable: true
                    description: 'Contains error message if any'
                required: 
                  - 'code'
  /auth/refreshToken: 
    post: 
      tags: 
        - 'Auth'
      operationId: 'refreshToken'
      requestBody: 
        content: 
          application/json: 
            schema: 
              type: 'object'
              properties: 
                refreshToken: 
                  type: 'string'
      responses: 
        200: 
          description: 'Success'
          content: 
            application/json: 
              schema: 
                type: 'object'
                properties: 
                  code: 
                    type: 'integer'
                    description: 'The status code of the response'
                  result: 
                    $ref: '#/components/schemas/AuthToken'
                    nullable: false
                    description: ''
                  error: 
                    type: 'string'
                    nullable: true
                    description: 'Contains error message if any'
                required: 
                  - 'code'
...
components: 
  schemas: 
...
    AuthToken: 
      type: 'object'
      description: ''
      properties: 
        accessToken: 
          type: 'string'
          nullable: false
          description: ''
        refreshToken: 
          type: 'string'
          nullable: false
          description: ''
        expiresAt: 
          type: 'string'
          nullable: false
          description: ''
      required: 
        - 'accessToken'
        - 'refreshToken'
        - 'expiresAt'
    User: 
      type: 'object'
      description: ''
      properties: 
        id: 
          type: 'string'
          nullable: false
          description: ''
        email: 
          type: 'string'
          nullable: false
          description: ''
        name: 
          type: 'string'
          nullable: false
          description: ''
        profilePicture: 
          type: 'string'
          nullable: true
          description: ''
        phoneNumber: 
          type: 'string'
          nullable: true
          description: ''
        username: 
          type: 'string'
          nullable: true
          description: ''
        multiFactorAuthEnabled: 
          type: 'boolean'
          nullable: false
          description: ''
        selectedCompanyId: 
          type: 'string'
          nullable: true
          description: ''
        status: 
          $ref: '#/components/schemas/AccountStatus'
          nullable: true
          description: ''
        fcmTokens: 
          type: 'array'
          nullable: false
          description: ''
          items: 
            type: 'array'
            nullable: false
            description: ''
      required: 
        - 'id'
        - 'email'
        - 'name'
        - 'multiFactorAuthEnabled'
        - 'fcmTokens'
    AccountStatus: 
      type: 'string'
      enum: 
        - 'CHANGE_PASSWORD'
        - 'COMPLETED'
...
    UserUpdateRequest: 
      type: 'object'
      description: ''
      properties: 
        name: 
          type: 'string'
          nullable: true
          description: ''
        profilePicture: 
          type: 'string'
          nullable: true
          description: ''
        phoneNumber: 
          type: 'string'
          nullable: true
          description: ''
        fcmTokens: 
          type: 'array'
          nullable: true
          description: ''
          items: 
            type: 'array'
            nullable: true
            description: ''
...